Refactor (#28)
