+++ src/Responses/NotificationResponse.php
@@ -2,15 +2,13 @@
 
 namespace Coreproc\PaynamicsSdk\Responses;
 
+use Coreproc\PaynamicsSdk\Responses\Traits\GatewayResponse;
 use Coreproc\PaynamicsSdk\Responses\Traits\Interpreter;
 
 class NotificationResponse
 {
-    use Interpreter;
+    use Interpreter, GatewayResponse;
 
-    /**
-     * @var string
-     */
     public string $response;
 
     /**
@@ -21,88 +19,6 @@ class NotificationResponse
     public static function make(): NotificationResponse
     {
         return new self();
-    }
-
-    /**
-     * Set response of notification
-     *
-     * @param string $response
-     * @return $this
-     */
-    public function setResponse(string $response): NotificationResponse
-    {
-        $this->response = $response;
-        return $this;
-    }
-
-    /**
-     * Get merchant ID
-     *
-     * @return mixed
-     */
-    public function merchantId()
-    {
-        return $this->toArray()['merchantid'];
-    }
-
-    /**
-     * Get request id
-     *
-     * @return mixed
-     */
-    public function requestId()
-    {
-        return $this->toArray()['request_id'];
-    }
-
-    /**
-     * Get response id
-     *
-     * @return mixed
-     */
-    public function responseId()
-    {
-        return $this->toArray()['response_id'];
-    }
-
-    /**
-     * Get response code
-     *
-     * @return mixed
-     */
-    public function responseCode()
-    {
-        return $this->toArray()['response_code'];
-    }
-
-    /**
-     * Get response message
-     *
-     * @return mixed
-     */
-    public function responseMessage()
-    {
-        return $this->toArray()['response_message'];
-    }
-
-    /**
-     * Get response advise
-     *
-     * @return mixed
-     */
-    public function responseAdvise()
-    {
-        return $this->toArray()['response_advise'];
-    }
-
-    /**
-     * Get timestamp
-     *
-     * @return mixed
-     */
-    public function timeStamp()
-    {
-        return $this->toArray()['timestamp'];
     }
 
     /**
@@ -146,26 +62,6 @@ class NotificationResponse
     }
 
     /**
-     * Get processor response id
-     *
-     * @return mixed
-     */
-    public function processorResponseId()
-    {
-        return $this->toArray()['processor_response_id'];
-    }
-
-    /**
-     * Get processor response auth code
-     *
-     * @return mixed
-     */
-    public function processorResponseAuthCode()
-    {
-        return $this->toArray()['processor_response_authcode'];
-    }
-
-    /**
      * Get settlement info details
      *
      * @return mixed
@@ -173,15 +69,5 @@ class NotificationResponse
     public function settlementInfoDetails()
     {
         return $this->toArray()['settlement_info_details'];
-    }
-
-    /**
-     * Get signature
-     *
-     * @return mixed
-     */
-    public function signature()
-    {
-        return $this->toArray()['signature'];
     }
 }
